Fix HMD tracking; Include Library Folder to reduce incompatibility and fix crashes.
Turns out you need a [CameraRig], even if it isn't being used for
drawing, or else you cannot properly track the camera, so overlays are
stuck to the screen instead of part of the world. I've included the
Library Folder because I found Unity crashed a lot when it wasn't
included, even though Unity rebuilds it when it isn't found, it
apparently doesn't rebuild it properly. Unity does not crash for me
running the original version of the Overlay project, but without the
Library folder it crashes within about 2 minutes. This was built using
Unity 5.3.5f1.