Fixing SteamVR
- Stop forcing Native VR Support automatically because Unity does not
support Native VR Applications which do not draw a scene.
- Stop deleting the OpenVR dlls since they are needed if we aren't
building with Native VR Support, which we cannot do unless we are
drawing a scene.
- Fix SteamVR.cs to allow Applications which do not draw a scene.
- Adds back the default SteamVR_Overlay.shader - not sure why this is
missing?
- Adds back the missing OpenVR dlls since they are needed for builds
without Native VR Support.

This seems to be the minimum that is required for Unity 5.6.0f3 to work
properly with SteamVR without Native VR Support enabled in Unity. This
is required when we are not drawing a scene, so regular scene
applications which are using HOTK will not need to make the changes in
this commit.

Aside from the .meta files, only 3 files were added and only 3 files
changed in this commit, so I think the impact is about as small as it
can get. Will have to finish testing with other HMDs to ensure these
changes (specifically the ones to SteamVR.cs) are working 100%.